--- app/src/main/java/com/bscpe3g/sugboeats/MyReservationScreen.java
@@ -1,62 +1,0 @@
-package com.bscpe3g.sugboeats;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.ImageButton;
-
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
-public class MyReservationScreen extends AppCompatActivity {
-
-    ImageButton home, sched, fav, prof;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_my_reservation_screen);
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.my_reservation_screen), (v, insets) -> {
-            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
-            return insets;
-        });
-
-        home = findViewById(R.id.imagehome);
-        sched = findViewById(R.id.imagesched);
-        fav = findViewById(R.id.imagefave);
-        prof = findViewById(R.id.imageprof);
-
-        home.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(MyReservationScreen.this, ResetLink.class));
-            }
-        });
-
-        sched.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(MyReservationScreen.this, MyReservationScreen.class));
-            }
-        });
-
-        fav.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(MyReservationScreen.this, ResetLink.class));
-            }
-        });
-
-        prof.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(MyReservationScreen.this, ResetLink.class));
-            }
-        });
-    }
-}
