+++ app/src/main/res/layout/cardview.xml
@@ -1,0 +1,143 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    app:cardElevation="8dp"
+    app:cardCornerRadius="10dp"
+    android:layout_margin="16dp"
+    >
+
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <View
+                android:id="@+id/view7"
+                android:layout_width="360dp"
+                android:layout_height="135dp"
+                android:layout_marginTop="16dp"
+                android:background="@drawable/reserve_viewer"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintHorizontal_bias="0.473"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <TextView
+                android:id="@+id/resto_name"
+                android:layout_width="160dp"
+                android:layout_height="20dp"
+                android:layout_marginStart="24dp"
+                android:layout_marginTop="14dp"
+                android:text="Sugbusog Grill House"
+                style="@style/resto_name"
+                app:layout_constraintStart_toStartOf="@+id/view7"
+                app:layout_constraintTop_toTopOf="@+id/view7" />
+
+            <TextView
+                android:id="@+id/date_time"
+                android:layout_width="160dp"
+                android:layout_height="18dp"
+                android:layout_marginTop="1dp"
+                android:text="49173947135"
+                style="@style/date_and_time"
+                app:layout_constraintEnd_toEndOf="@+id/resto_name"
+                app:layout_constraintStart_toStartOf="@+id/resto_name"
+                app:layout_constraintTop_toBottomOf="@+id/resto_name" />
+
+            <TextView
+                android:id="@+id/resto_loc"
+                android:layout_width="160dp"
+                android:layout_height="18dp"
+                android:layout_marginTop="1dp"
+                android:text="Palnsdnambfaf"
+                style="@style/resto_loc"
+                app:layout_constraintEnd_toEndOf="@+id/date_time"
+                app:layout_constraintStart_toStartOf="@+id/date_time"
+                app:layout_constraintTop_toBottomOf="@+id/date_time" />
+
+            <TextView
+                android:id="@+id/textView28"
+                android:layout_width="wrap_content"
+                android:layout_height="16dp"
+                android:layout_marginTop="1dp"
+                android:fontFamily="sans-serif"
+                android:text="Setting Type:"
+                android:textColor="@color/black"
+                android:textSize="12sp"
+                android:textStyle="bold"
+                app:layout_constraintStart_toStartOf="@+id/resto_loc"
+                app:layout_constraintTop_toBottomOf="@+id/resto_loc" />
+
+            <TextView
+                android:id="@+id/set_type"
+                style="@style/set_type"
+                android:layout_width="wrap_content"
+                android:layout_height="16dp"
+                android:layout_marginTop="1dp"
+                android:paddingStart="2dp"
+                android:text="Alfresco"
+                app:layout_constraintEnd_toEndOf="@+id/resto_loc"
+                app:layout_constraintHorizontal_bias="0.0"
+                app:layout_constraintStart_toEndOf="@+id/textView28"
+                app:layout_constraintTop_toBottomOf="@+id/resto_loc" />
+
+            <TextView
+                android:id="@+id/textView31"
+                android:layout_width="wrap_content"
+                android:layout_height="16dp"
+                android:layout_marginTop="1dp"
+                android:fontFamily="sans-serif"
+                android:text="Number of Guests:"
+                android:textColor="@color/black"
+                android:textSize="12sp"
+                android:textStyle="bold"
+                app:layout_constraintStart_toStartOf="@+id/textView28"
+                app:layout_constraintTop_toBottomOf="@+id/textView28" />
+
+            <TextView
+                android:id="@+id/num_guest"
+                style="@style/num_guest"
+                android:layout_width="wrap_content"
+                android:layout_height="16dp"
+                android:layout_marginTop="1dp"
+                android:text="10"
+                app:layout_constraintEnd_toEndOf="@+id/set_type"
+                app:layout_constraintHorizontal_bias="1.0"
+                app:layout_constraintStart_toEndOf="@+id/textView31"
+                app:layout_constraintTop_toBottomOf="@+id/set_type" />
+
+            <ImageView
+                android:id="@+id/imageView2"
+                android:layout_width="124dp"
+                android:layout_height="102dp"
+                android:layout_marginTop="12dp"
+                android:layout_marginEnd="20dp"
+                app:layout_constraintEnd_toEndOf="@+id/view7"
+                app:layout_constraintTop_toTopOf="@+id/view7"
+                app:srcCompat="@drawable/alfresco" />
+
+            <TextView
+                android:id="@+id/decision"
+                style="@style/decision"
+                android:layout_width="wrap_content"
+                android:layout_height="16dp"
+                android:text="Cancel Reservation"
+                android:textColor="@color/red"
+                app:layout_constraintBottom_toBottomOf="@+id/view7"
+                app:layout_constraintEnd_toEndOf="@+id/view7"
+                app:layout_constraintHorizontal_bias="0.501"
+                app:layout_constraintStart_toStartOf="@+id/view7"
+                app:layout_constraintTop_toTopOf="@+id/view7"
+                app:layout_constraintVertical_bias="0.957" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </LinearLayout>
+</androidx.cardview.widget.CardView>
