+++ app/src/main/res/values/colors.xml
@@ -5,4 +5,6 @@
     <color name="blue">#3282B8</color>
     <color name="grey">#5F6368</color>
     <color name="light_blue">#BBE1FA</color>
+    <color name="dark_blue">#1B262C</color>
+    <color name="red">#F90000</color>
 </resources>
