+++ app/src/main/res/values/strings.xml
@@ -93,6 +93,5 @@
         <item name="android:fontFamily">sans-serif</item>
     </style>
 
-    
 
 </resources>
